% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{get_all_dataelements}
\alias{get_all_dataelements}
\title{Retrieve all data elements in a DHIS2 Traker instance}
\usage{
get_all_dataelements()
}
\description{
Retrieve all data elements in a DHIS2 Traker instance
}
