% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{get_program_attributes}
\alias{get_program_attributes}
\title{Retrieve all programs attributes}
\usage{
get_program_attributes(program_name = NULL, program_id = NULL)
}
\description{
Retrieve all programs attributes
}
