% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{get_program_data_elements}
\alias{get_program_data_elements}
\title{Retrieve all data elements of a specific program}
\usage{
get_program_data_elements(program_name = NULL, program_id = NULL)
}
\description{
Retrieve all data elements of a specific program
}
