% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_data.R
\name{get_program_events_data}
\alias{get_program_events_data}
\title{Retrieve events data for a specific program}
\usage{
get_program_events_data(program_name = NULL, program_id = NULL,
  ou = NULL)
}
\description{
Retrieve events data for a specific program
}
