% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{get_program_id}
\alias{get_program_id}
\title{Retrieve the program id  of a program in a DHIS2 Tracker instance}
\usage{
get_program_id(program_name)
}
\description{
Retrieve the program id  of a program in a DHIS2 Tracker instance
}
