% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdhis2tracker.R
\docType{package}
\name{rdhis2tracker}
\alias{rdhis2tracker}
\alias{Rdhis2tracker}
\alias{Rdhis2tracker-package}
\alias{rdhis2tracker-package}
\title{DHIS2 Tracker API client library}
\description{
A package connecting R to the DHIS2 Tracker API
}
