% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{get_program_orgunits}
\alias{get_program_orgunits}
\title{Retrieve all orgs units of a specific program}
\usage{
get_program_orgunits(program_name = NULL, program_id = NULL)
}
\description{
Retrieve all orgs units of a specific program
}
