% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_data.R
\name{get_program_tracked_entities_data}
\alias{get_program_tracked_entities_data}
\title{Retrieve entities data for a specific program}
\usage{
get_program_tracked_entities_data(program_name = NULL,
  program_id = NULL, ou = NULL)
}
\arguments{
\item{program_name}{program name to retreive the data}

\item{program_id}{uid of the program to retreive the data}

\item{ou}{organisation unit uid of which you want to retreive the data}
}
\description{
Retrieve entities data for a specific program
}
